update PacksList component and add PackList component
