+++ src/components/Container/Main/PacksListContainer/PacksListContainer.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import style from './PacksListContainer.module.css'
 import {Redirect} from "../../../../common/Redirect/Redirect";
 import {CardsFilter} from "../Profile/CardsFilter/CardsFilter";
+import {PackList} from "./PacksList/PackList/PackList";
 import {PacksList} from "./PacksList/PacksList";
 
 type ProfilePropsType = {
@@ -10,17 +11,12 @@ type ProfilePropsType = {
 
 const PacksListContainer: React.FC<ProfilePropsType> = () => {
     return (
-        <div className={style.profile__container}>
-            <div className={style.profile__body}>
-                <div className={style.profile__body_profile}>
-
-
-                    {/*Компонента с отрисовкой фильра по количеству карт*/}
+        <div className={style.packsList__container}>
+            <div className={style.packsList__body}>
+                <div className={style.packsList__body_profile}>
                     <CardsFilter/>
                 </div>
-                <div className={style.profile__body_main}>
-
-                    {/*Компонента с отрисовкой карточек*/}
+                <div className={style.packsList__body_main}>
                     <PacksList/>
                 </div>
             </div>
